backend/fastapi_app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg2-binary
pydantic
python-dotenv
stripe
sendgrid
google-cloud-storage
celery
redis